chore: Update color and font dependencies
The color.dart file was modified to add a new color constant for white text. Additionally, the pubspec.yaml file was updated to include the google_fonts package as a dependency. The main.dart file was also modified to set the system UI overlay style. Finally, the loading_page.dart file was updated to add a delay before navigating to the WelcomePage.

```